Interfaz Lección y Niveles v0.1
se agregaron nuevas screen y un poco de codigo y estilo en cada una como el home y lecciones, falta agregar mas screens y estilos por eso es versine0.1 hasta llegar a la version 1
