--- Triolingo-app/src/components/Main.js
@@ -1,19 +1,0 @@
-import React from 'react';
-import { NavigationContainer } from '@react-navigation/native';
-import { createDrawerNavigator } from '@react-navigation/drawer';
-import HomeScreen from './HomeScreen';
-import LeccionScreen from '../screens/Lecciones';
-const Drawer = createDrawerNavigator();
-
-const App = () => {
-  return (
-    <NavigationContainer>
-      <Drawer.Navigator initialRouteName="Home">
-        <Drawer.Screen name="Home" component={HomeScreen} />
-        <Drawer.Screen name="Lecciones" component={LeccionScreen} />
-      </Drawer.Navigator>
-    </NavigationContainer>
-  );
-};
-
-export default App;
