--- Triolingo-app/src/screens/Nav-screen.js
@@ -1,19 +1,0 @@
-// import React from 'react';
-// import { View, Text } from 'react-native';
-// import NavigationBar from './components/Nav-bar';
-// const MyScreen = () => {
-
-//   return (
-//     <View>
-//       <NavigationBar
-//         title="TRIOLINGO"
-//         onLeftButtonPress={handleLeftButtonPress}
-//         leftButtonTitle="Izquierda"
-//         onRightButtonPress={handleRightButtonPress}
-//         rightButtonTitle="Derecha"
-//       />
-//     </View>
-//   );
-// };
-
-// export default MyScreen;
